RecyclerView

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,12 +1,11 @@
 package ru.netology.nmedia.dto
 
 data class Post(
-    val id: Int,
+    val id: Long,
     val author: String,
     val content: String,
     val published: String,
     val likedByMe: Boolean = false,
     val likesCount: Int = 0,
     val shareCount: Int = 0
-
 )
